pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "ai-meeting-room"
version = "0.1.0"
description = "Multi-AI Meeting Room with Poe API"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.85.0",
    "openai>=1.50.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
